requirements.txt
asgiref==3.2.10
Django==3.1.2
django-bootstrap-modal-forms==2.0.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
djangorestframework==3.12.1
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
